requirements.txt
Flask
PyPDF2
python-dotenv
google-generativeai